+++ jwst/assign_wcs/miri.py
@@ -11,7 +11,8 @@ from gwcs import selector
 from gwcs.utils import _toindex
 from . import pointing
 from ..transforms import models as jwmodels
-from .util import not_implemented_mode, subarray_transform
+from .util import (not_implemented_mode, subarray_transform,
+                   velocity_correction)
 from ..datamodels import (DistortionModel, FilteroffsetModel,
                           DistortionMRSModel, WavelengthrangeModel,
                           RegionsModel, SpecwcsModel)
@@ -37,7 +38,9 @@ def create_pipeline(input_model, reference_files):
     """
     exp_type = input_model.meta.exposure.type.lower()
     pipeline = exp_type2transform[exp_type](input_model, reference_files)
-
+    if pipeline:
+        log.info("Created a MIRI {0} pipeline with references {1}".format(
+            exp_type, reference_files))
     return pipeline
 
 
@@ -190,6 +193,16 @@ def lrs(input_model, reference_files):
     # Create the model transforms.
     lrs_wav_model = jwmodels.LRSWavelength(lrsdata, zero_point)
 
+    try:
+        velosys = input_model.meta.wcsinfo.velosys
+    except AttributeError:
+        pass
+    else:
+        if velosys is not None:
+            velocity_corr = velocity_correction(input_model.meta.wcsinfo.velosys)
+            lrs_wav_model = lrs_wav_model | velocity_corr
+            log.info("Applied Barycentric velocity correction : {}".format(velocity_corr[1].amplitude.value))
+
     # Incorporate the small rotation
     angle = np.arctan(0.00421924)
     rot = models.Rotation2D(angle)
@@ -237,6 +250,7 @@ def ifu(input_model, reference_files):
     det2abl = (detector_to_abl(input_model, reference_files)).rename(
         "detector_to_abl")
     abl2v2v3l = (abl_to_v2v3l(input_model, reference_files)).rename("abl_to_v2v3l")
+
     tel2sky = pointing.v23tosky(input_model) & models.Identity(1)
 
     # Put the transforms together into a single transform
@@ -280,6 +294,16 @@ def detector_to_abl(input_model, reference_files):
 
     with SpecwcsModel(reference_files['specwcs']) as f:
         lambda_model = f.model
+
+    try:
+        velosys = input_model.meta.wcsinfo.velosys
+    except AttributeError:
+        pass
+    else:
+        if velosys is not None:
+            velocity_corr = velocity_correction(input_model.meta.wcsinfo.velosys)
+            lambda_model = [m | velocity_corr for m in lambda_model]
+            log.info("Applied Barycentric velocity correction : {}".format(velocity_corr[1].amplitude.value))
 
     with RegionsModel(reference_files['regions']) as f:
         regions = f.regions.copy()